Add files via upload
Error handling is a fundamental python concept. 
Exceptions are used to report errors when are commited by external users. Some fundamental functions that may help U are try-except and raise. 
Assertions are statements that must be true otherwise it should be some bug in the code.
In this folder U can view two examples of exceptions and assertions used to solve the problem that appears when users write undesireable inputs (In this case chars and negative integers).